added model relationships and access control

+++ app/Http/Controllers/PostsController.php
@@ -8,14 +8,21 @@ use App\Post;
 class PostsController extends Controller
 {
     /**
+     * Create a new controller instance.
+     *
+     * @return void
+     */
+    public function __construct()
+    {
+        $this->middleware('auth', ['except'=> ['index', 'show']]);
+    }
+    /**
      * Display a listing of the resource.
      *
      * @return \Illuminate\Http\Response
      */
     public function index()
     {
-        //$posts = Post::orderBy('created_at','asc')->get();
-        //$posts = Post::orderBy('created_at','asc')->take(10)->get();
         $posts = Post::orderBy('created_at','dsc')->paginate(5);
         return view('posts.index')->with('posts', $posts);
     }
@@ -75,6 +82,10 @@ class PostsController extends Controller
     public function edit($id)
     {
         $post = Post::find($id);
+        //check authorized user
+        if(auth()->user()->id !==$post->user_id){
+            return redirect('/posts')->with('error','Unauthorized Access');
+        }
         return view('posts.edit')->with('post', $post);
     }
 
@@ -111,6 +122,10 @@ class PostsController extends Controller
     public function destroy($id)
     {
         $post = Post::find($id);
+        //check authorized user
+        if(auth()->user()->id !==$post->user_id){
+            return redirect('/posts')->with('error','Unauthorized Access');
+        }
         $post->delete();
         return redirect('/posts')->with('success','Post Removed');
     }
